Add methods to RectPrimitive
