+++ sketchlib/Rectangle.js
@@ -17,45 +17,4 @@ export class Rectangle {
     this.position = position;
     this.dimensions = dimensions;
   }
-
-  /**
-   * The center of the rectangle
-   * @type {Point}
-   */
-  get center() {
-    return this.position.add(this.dimensions.scale(0.5));
-  }
-
-  /**
-   * Returns true if the point is inside the rectangle
-   * @param {Point} point A point to compare with the rectangle
-   * @returns {boolean} True if the point is within the bounds of the rectangle
-   */
-  contains(point) {
-    const { x, y } = point;
-    const { x: x_min, y: y_min } = this.position;
-    const { x: width, y: height } = this.dimensions;
-    const x_max = x_min + width;
-    const y_max = y_min + height;
-    return x >= x_min && x < x_max && y >= y_min && y < y_max;
-  }
-
-  /**
-   * Compute a rectangle from a center point and the dimensions
-   * @param {Point} center The center of the
-   * @param {Direction} dimensions
-   * @returns {Rectangle}
-   */
-  static from_center(center, dimensions) {
-    return new Rectangle(center.add(dimensions.scale(-0.5)), dimensions);
-  }
 }
-
-/**
- * A rectangle that sets the boundary of the screen
- * @type {Rectangle}
- */
-export const SCREEN_RECT = new Rectangle(
-  Point.ORIGIN,
-  new Direction(WIDTH, HEIGHT),
-);
